complete project before delete previous design css

+++ src/context/ProductContextData.jsx
@@ -47,38 +47,40 @@ export default function ProductContextData({ children }) {
   const [productUpdateResponse, setProductUpdateResponse] = useState(false);
 
   const getProducts = async () => {
-    try {
-      setIsLoading(true);
-      const response = await getAllProducts(
-        pageNumber,
-        pageSize,
-        searchValue,
-        sort
-      );
-      // console.log("***********from product services*************");
-      // console.log(response);
-      const data = response.data;
-      // console.log("data=" + data);
-      const productsData = data.items.$values;
-      // console.log("productsData" + productsData);
-      setProducts(productsData);
-      // setTotalPages(data.totalItems);
-      // const totalItems = data.items.$values.length;
-      const totalItems = data.totalItems;
-      // console.log("totalItems=> " + totalItems + " | pageSize=>" + pageSize);
-      // console.log("totalItems / pageSize =" + Math.ceil(totalItems / pageSize));
-      setTotalPages(Math.ceil(totalItems / pageSize));
-      setData(data);
-    } catch (error) {
-      setError(error);
-    } finally {
-      setIsLoading(false);
-    }
+      try {
+        setIsLoading(true);
+        const response = await getAllProducts(
+          pageNumber,
+          pageSize,
+          searchValue,
+          sort
+        );
+        // console.log("***********from product services*************");
+        // console.log(response);
+        const data = response.data;
+        // console.log("data=" + data);
+        const productsData = data.items.$values;
+        // console.log("productsData" + productsData);
+        setProducts(productsData);
+        // setTotalPages(data.totalItems);
+        // const totalItems = data.items.$values.length;
+        const totalItems = data.totalItems;
+        // console.log("totalItems=> " + totalItems + " | pageSize=>" + pageSize);
+        // console.log("totalItems / pageSize =" + Math.ceil(totalItems / pageSize));
+        setTotalPages(Math.ceil(totalItems / pageSize));
+        setData(data);
+      } catch (error) {
+        setError(error);
+      } finally {
+        setIsLoading(false);
+      }
   };
 
   // create product
   const addNewProduct = async () => {
     console.log("addNewProduct");
+        if ( !createProduct) return;
+
     try {
       setIsLoading(true);
       // call product create product service
@@ -111,30 +113,36 @@ export default function ProductContextData({ children }) {
   // delete product
   const deleteProduct = async () => {
     console.log("**deleteProduct response");
-    try {
-      setIsLoading(true);
-      const response = await deleteProductByID(deleteProductID, adminToken);
-      console.log(" deleteProduct response", response);
-      setProductDeleteResponse(true);
-    } catch (error) {
-      setError(error);
-    } finally {
-      setIsLoading(false);
-    }
+    if (!deleteProductID) return;
+      try {
+        setIsLoading(true);
+        const response = await deleteProductByID(deleteProductID, adminToken);
+        console.log(" deleteProduct response", response);
+        setProductDeleteResponse(true);
+      } catch (error) {
+        setError(error);
+      } finally {
+        setIsLoading(false);
+      }
   };
 
   // update Product
   const updateProductByID = async () => {
-    try {
-      setIsLoading(true);
-      console.log("updateProductByID=>", productData, productID, adminToken);
-      const response = await updateProduct(productData, productID, adminToken);
-      console.log(response);
-    } catch (error) {
-      setError(error);
-    } finally {
-      setIsLoading(false);
-    }
+    if (!productData|| !productID || !adminToken) return;
+      try {
+        setIsLoading(true);
+        console.log("updateProductByID=>", productData, productID, adminToken);
+        const response = await updateProduct(
+          productData,
+          productID,
+          adminToken
+        );
+        console.log(response);
+      } catch (error) {
+        setError(error);
+      } finally {
+        setIsLoading(false);
+      }
   };
 
   // useEffect
@@ -202,6 +210,7 @@ export default function ProductContextData({ children }) {
           setUpdateProductOrder,
           productUpdateResponse,
           setProductUpdateResponse,
+   
         }}
       >
         {children}
